refactor: `resample.py` -> `utils/xarray.py` and `utils/gdal_formats.py`
