+++ python/pyproject.toml
@@ -77,8 +77,7 @@ addopts = """
     --ignore=clim_recal/debiasing/python-cmethods
     --ignore=clim_recal/debiasing/preprocess_data.py
     --ignore=clim_recal/debiasing/run_cmethods.py
-    -m "not server"
-    -m "not mount"
+    -m "not server and not mount"
     --pdbcls=IPython.terminal.debugger:TerminalPdb
     --cov=clim_recal
     --cov-report=term:skip-covered
